refactor(database): move different type of database queries/actions to different files, ie guestbook, comments, and views
