+++ src/lib/auth.ts
@@ -1,4 +1,4 @@
-import NextAuth from 'next-auth';
+import NextAuth, { Session } from 'next-auth';
 import GitHubProvider from 'next-auth/providers/github';
 import DiscordProvider from 'next-auth/providers/discord';
 import SpotifyProvider from 'next-auth/providers/spotify';
@@ -22,3 +22,13 @@ export const {
     }),
   ],
 });
+
+export async function getSession(): Promise<Session> {
+  let session = await auth();
+  
+  if (!session || !session.user) {
+    throw new Error('Unauthorized');
+  }
+
+  return session;
+}
