+++ src/lib/database-queries/views.ts
@@ -47,49 +47,3 @@ export async function getViewCount(slug: string): Promise<
     WHERE slug=(${slug});
   `;
 }
-
-export async function getGuestbookEntries(): Promise<
-  EntryData[]
-> {
-  if (!process.env.POSTGRES_URL) {
-    return [];
-  }
-
-  return sql`
-    SELECT id, body, created_by, created_at, updated_at, email, color
-    FROM guestbook
-    ORDER BY created_at DESC
-    LIMIT 300;
-  `;
-}
-
-export async function getComments({ slug }: { slug: string }): Promise<
-  CommentData[]
-> {
-  if (!process.env.POSTGRES_URL) {
-    return [];
-  }
-  
-  return sql`
-    SELECT id, body, created_by, created_at, updated_at, email
-    FROM comments
-    WHERE slug = (${slug})
-    ORDER BY created_at DESC
-    LIMIT 15;
-  `;
-}
-
-export async function getEveryComment(): Promise<
-  CommentData[]
-> {
-  if (!process.env.POSTGRES_URL) {
-    return [];
-  }
-  
-  return sql`
-    SELECT id, slug, body, created_by, created_at, updated_at, email
-    FROM comments
-    ORDER BY created_at DESC
-    LIMIT 15;
-  `;
-}
